C-API
