+++ machine/capi/capi.hpp
@@ -17,8 +17,8 @@
 #include "capi/capi_constants.h"
 #include "capi/tag.hpp"
 
-#define ENTER_CAPI(state) (state->vm()->shared.enter_capi(state, __FILE__, __LINE__))
-#define LEAVE_CAPI(state) (state->vm()->shared.leave_capi(state))
+#define ENTER_CAPI(state) (state->c_api()->enter_capi(state, __FILE__, __LINE__))
+#define LEAVE_CAPI(state) (state->c_api()->leave_capi(state))
 
 namespace rubinius {
   class Integer;
